attempt to add tests in seperate files
